Render line preview
