+++ docs/data/charts/zoom-and-pan/ZoomSliderPreview.tsx.preview
@@ -1,0 +1,10 @@
+<Typography variant="h6" sx={{ alignSelf: 'center' }}>
+  Unemployment Rate in United States (1948-2025)
+</Typography>
+<LineChartPro
+  {...settings}
+  xAxis={[{ ...xAxis, zoom: { slider: { enabled: true, preview: true } } }]}
+/>
+<Typography variant="caption">
+  Source: Federal Reserve Bank of St. Louis. Updated: Jun 6, 2025 7:46 AM CDT.
+</Typography>
